requirements.txt
asgiref==3.8.1
Django==4.2.6
django-ckeditor==6.7.0
django-js-asset==2.2.0
pillow==10.4.0
PyMySQL==1.1.0
PyYAML==6.0.1
sqlparse==0.5.1
